added replies and media filters in userProfile, and refactored some of the component. #71
